first steps to prepare for PyPI
