+++ sntools/detectors.py
