+++ sntools/formats/__init__.py
