+++ sntools/formats/gamma.py
