+++ sntools/formats/princeton.py
