+++ sntools/interaction_channels/c12e.py
@@ -1,32 +1,37 @@
 '''
-Implementation of nu_e + 16O -> X + e-
+Implementation of nu_e + 12C -> X + e-
 
-Based on Appendix B.3 of hep-ph/0307050.
-That paper only gives the total cross section sigma(eNu), not the differential
-c.s. dSigma/dE (eNu, eE), but since we assume eE = eNu - 15 MeV, we can write
-the differential c.s. as sigma(eNu) * delta(eNu - 15 - eE).
-However, numpy doesn't implement a delta distribution and numpy's (numerical)
-integration doesn't play nice with sympy's (symbolic) DiracDelta, see:
-https://stackoverflow.com/questions/36755487/diracdelta-not-giving-correct-result#36755974
-Instead, below we implement an approximation to DiracDelta: a function that's
+Based on arXiv:nucl-th/9903022, Table V. Following arXiv:1204.4231 (eqn. III.1),
+we fit a power series to the tabulated values. Furthermore, we ignore excited
+states of the final nucleus, so that the observed energy is eE = eNu - e_thr.
+
+To determine the the differential cross section dSigma/dE (eNu, eE) from the
+total cross section, we approximate a DiracDelta function with one that is
 2*epsilon wide and 1/(2*epsilon) high, so that the integral is 1.
 '''
 
-from event import Event
+from sntools.event import Event
+
+e_thr = 17.338 # energy threshold of this reaction (arXiv:nucl-th/0001050, p. 7)
+epsilon = 0.001 # for approximating DiracDelta distribution below
 
 
+'''
+generate_event(eNu, direction):
+Generate an event with appropriate incoming/outgoing particles.
+Input:
+    eNu: neutrino energy
+    dirx, diry, dirz: direction of outgoing particle (normalized to 1)
+Output:
+    event object
+'''
 def generate_event(eNu, dirx, diry, dirz):
     eE = get_eE(eNu, dirz)
-
-    evt = Event(-1008016)
-    evt.incoming_particles.append((-12, eNu, 0, 0, 1))  # incoming neutrino
-    evt.incoming_particles.append((8016, 14900, 0, 0, 1))  # oxygen nucleus at rest
-    evt.outgoing_particles.append((-11, eE, dirx, diry, dirz))  # outgoing positron
+    evt = Event(1006012)
+    evt.incoming_particles.append([12, eNu, 0, 0, 1]) # incoming nu_e
+    evt.incoming_particles.append((6012, 11178, 0, 0, 1)) # carbon nucleus at rest
+    evt.outgoing_particles.append([11, eE, dirx, diry, dirz]) # outgoing electron
     return evt
-
-
-e_thr = 15  # energy threshold for this reaction
-epsilon = 0.001  # for approximating DiracDelta distribution below
 
 
 '''
@@ -80,14 +85,15 @@ Input:
 Output:
     one floating point number
 '''
-def dSigma_dE(eNu, eE): # eq. (B6)
+def dSigma_dE(eNu, eE):
     if abs(get_eE(eNu) - eE) > epsilon:
-        # This should never be called since we set bounds_eE() accordingly above
+        # This should never happen, since we set bounds_eE() accordingly above
         # ... but just in case:
         return 0
 
-    sigma0 = 4.7E-40 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
-    sigma = sigma0 * (eNu**0.25 - 15**0.25)**6
+    sigma0 = 3.439E-44 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
+    a1, a2, a3 = 10.164, -0.4666, 0.0546
+    sigma = sigma0 * (a1*(eNu-e_thr) + a2*(eNu-e_thr)**2 + a3*(eNu-e_thr)**3)
     return sigma / (2*epsilon) # Ensure that integration over eE yields sigma
 
 
@@ -100,9 +106,12 @@ Input:
 Output:
     one floating point number
 '''
-def dSigma_dCosT(eNu, cosT): # eq. (B7)
-    x = (get_eE(eNu, cosT) / 25)**4
-    return 1 - cosT * (1+x)/(3+x)
+def dSigma_dCosT(eNu, cosT):
+    # Small values of cosT are preferred, see arXiv:hep-ex/0105068 (fig. 12,14).
+    # However, energy dependence is unclear, so we use a constant value for now.
+    if abs(cosT) > 1:
+        return 0
+    return 0.5
 
 
 # minimum/maximum neutrino energy that can produce a given positron energy
