+++ sntools/interaction_channels/c12eb.py
@@ -1,19 +1,18 @@
 '''
-Implementation of nu + 12C -> nu' + 12C*, 12C* -> 12C + \gamma
+Implementation of anti-nu_e + 12C -> X + e+
 
-Based on Donnelly & Peccei, Phys.Rept. 50 (1979) 1, eq. (4.53). We assume
-\beta * \kappa = 1.11, based on an experimental determination, see
-Armbruster et al. (KARMEN Collaboration), Phys.Lett.B423 (1998), p. 15.
-This is higher than the SM prediction (=1) but consistent with SNOwGLoBES.
+Based on arXiv:nucl-th/9903022, Table V. Following arXiv:1204.4231 (eqn. III.1),
+we fit a power series to the tabulated values. Furthermore, we ignore excited
+states of the final nucleus, so that the observed energy is eE = eNu - e_thr.
 
 To determine the the differential cross section dSigma/dE (eNu, eE) from the
 total cross section, we approximate a DiracDelta function with one that is
 2*epsilon wide and 1/(2*epsilon) high, so that the integral is 1.
 '''
 
-from event import Event
+from sntools.event import Event
 
-e_thr = 15.11 # energy threshold of this reaction
+e_thr = 14.39 # energy threshold of this reaction (arXiv:1507.05613, p. 74)
 epsilon = 0.001 # for approximating DiracDelta distribution below
 
 
@@ -27,14 +26,11 @@ Output:
     event object
 '''
 def generate_event(eNu, dirx, diry, dirz):
-    # Note: `builtins._flavor` is set in `genevts.py`
-    nu_flv = {'e':12, 'eb':-12, 'x':14, 'xb':-14}[_flavor]
-
-    evt = Event(2006012 if nu_flv > 0 else -2006012)
-    evt.incoming_particles.append([nu_flv, eNu, 0, 0, 1]) # incoming nu
+    eE = get_eE(eNu, dirz)
+    evt = Event(-1006012)
+    evt.incoming_particles.append([-12, eNu, 0, 0, 1]) # incoming neutrino
     evt.incoming_particles.append((6012, 11178, 0, 0, 1)) # carbon nucleus at rest
-    evt.outgoing_particles.append([22, e_thr, dirx, diry, dirz]) # emitted gamma
-#     evt.outgoing_particles.append([nu_flv, eNu-e_thr, 0, 0, 1]) # outgoing nu
+    evt.outgoing_particles.append([-11, eE, dirx, diry, dirz]) # outgoing positron
     return evt
 
 
@@ -42,7 +38,7 @@ def generate_event(eNu, dirx, diry, dirz):
 possible_flavors:
 which neutrino flavors ("e", "eb", "x", "xb") interact in this channel
 '''
-possible_flavors = ("e", "eb", "x", "xb")
+possible_flavors = ["eb"]
 
 
 '''
@@ -51,7 +47,7 @@ List with minimum & maximum energy of incoming neutrino. The minimum energy is
 typically given by the threshold energy for the interaction, while the maximum
 energy is given by the supernova neutrino flux.
 '''
-bounds_eNu = (e_thr, 100)
+bounds_eNu = [e_thr + 0.8, 100] # 0.8 MeV = Cherenkov threshold of positron
 
 
 '''
@@ -69,7 +65,7 @@ def bounds_eE(eNu, *args):
 
 '''
 get_eE(eNu, cosT):
-Energy of outgoing (detected) particle.
+Energy of outgoing (detected particle).
 Input:
     eNu:  neutrino energy (MeV)
     cosT: cosine of the angle between neutrino and outgoing (detected) particle
@@ -77,7 +73,7 @@ Output:
     one floating point number
 '''
 def get_eE(eNu, cosT=0):
-    return e_thr
+    return eNu - e_thr
 
 
 '''
@@ -90,14 +86,14 @@ Output:
     one floating point number
 '''
 def dSigma_dE(eNu, eE):
-    if eNu < e_thr or abs(get_eE(eNu) - eE) > epsilon:
-        # This should never happen, since we set bounds for eE and eNu accordingly above
+    if abs(get_eE(eNu) - eE) > epsilon:
+        # This should never happen, since we set bounds_eE() accordingly above
         # ... but just in case:
         return 0
 
-    sigma0 = 1.08E-38 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
-    sigma = sigma0 * ((eNu-e_thr) / 939)**2
-    sigma *= 1.11**2 # coupling constant \beta (see discussion above)
+    sigma0 = 9.55E-46 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
+    a1, a2, a3 = -106.3, 25.15, 0.3697
+    sigma = sigma0 * (a1*(eNu-e_thr) + a2*(eNu-e_thr)**2 + a3*(eNu-e_thr)**3)
     return sigma / (2*epsilon) # Ensure that integration over eE yields sigma
 
 
@@ -117,6 +113,6 @@ def dSigma_dCosT(eNu, cosT):
     return 0.5
 
 
-# min/max neutrino energy to produce outgoing particle with observed energy
+# minimum/maximum neutrino energy that can produce a given positron energy
 def _bounds_eNu(eE):
-    return bounds_eNu
+    return (eE + e_thr - epsilon, eE + e_thr + epsilon)
