+++ sntools/interaction_channels/c12nc.py
@@ -1,18 +1,19 @@
 '''
-Implementation of nu_e + 12C -> X + e-
+Implementation of nu + 12C -> nu' + 12C*, 12C* -> 12C + \gamma
 
-Based on arXiv:nucl-th/9903022, Table V. Following arXiv:1204.4231 (eqn. III.1),
-we fit a power series to the tabulated values. Furthermore, we ignore excited
-states of the final nucleus, so that the observed energy is eE = eNu - e_thr.
+Based on Donnelly & Peccei, Phys.Rept. 50 (1979) 1, eq. (4.53). We assume
+\beta * \kappa = 1.11, based on an experimental determination, see
+Armbruster et al. (KARMEN Collaboration), Phys.Lett.B423 (1998), p. 15.
+This is higher than the SM prediction (=1) but consistent with SNOwGLoBES.
 
 To determine the the differential cross section dSigma/dE (eNu, eE) from the
 total cross section, we approximate a DiracDelta function with one that is
 2*epsilon wide and 1/(2*epsilon) high, so that the integral is 1.
 '''
 
-from event import Event
+from sntools.event import Event
 
-e_thr = 17.338 # energy threshold of this reaction (arXiv:nucl-th/0001050, p. 7)
+e_thr = 15.11 # energy threshold of this reaction
 epsilon = 0.001 # for approximating DiracDelta distribution below
 
 
@@ -26,11 +27,14 @@ Output:
     event object
 '''
 def generate_event(eNu, dirx, diry, dirz):
-    eE = get_eE(eNu, dirz)
-    evt = Event(1006012)
-    evt.incoming_particles.append([12, eNu, 0, 0, 1]) # incoming nu_e
+    # Note: `builtins._flavor` is set in `genevts.py`
+    nu_flv = {'e':12, 'eb':-12, 'x':14, 'xb':-14}[_flavor]
+
+    evt = Event(2006012 if nu_flv > 0 else -2006012)
+    evt.incoming_particles.append([nu_flv, eNu, 0, 0, 1]) # incoming nu
     evt.incoming_particles.append((6012, 11178, 0, 0, 1)) # carbon nucleus at rest
-    evt.outgoing_particles.append([11, eE, dirx, diry, dirz]) # outgoing electron
+    evt.outgoing_particles.append([22, e_thr, dirx, diry, dirz]) # emitted gamma
+#     evt.outgoing_particles.append([nu_flv, eNu-e_thr, 0, 0, 1]) # outgoing nu
     return evt
 
 
@@ -38,7 +42,7 @@ def generate_event(eNu, dirx, diry, dirz):
 possible_flavors:
 which neutrino flavors ("e", "eb", "x", "xb") interact in this channel
 '''
-possible_flavors = ["e"]
+possible_flavors = ("e", "eb", "x", "xb")
 
 
 '''
@@ -47,7 +51,7 @@ List with minimum & maximum energy of incoming neutrino. The minimum energy is
 typically given by the threshold energy for the interaction, while the maximum
 energy is given by the supernova neutrino flux.
 '''
-bounds_eNu = [e_thr + 0.8, 100] # 0.8 MeV = Cherenkov threshold of electron
+bounds_eNu = (e_thr, 100)
 
 
 '''
@@ -65,7 +69,7 @@ def bounds_eE(eNu, *args):
 
 '''
 get_eE(eNu, cosT):
-Energy of outgoing (detected particle).
+Energy of outgoing (detected) particle.
 Input:
     eNu:  neutrino energy (MeV)
     cosT: cosine of the angle between neutrino and outgoing (detected) particle
@@ -73,7 +77,7 @@ Output:
     one floating point number
 '''
 def get_eE(eNu, cosT=0):
-    return eNu - e_thr
+    return e_thr
 
 
 '''
@@ -86,14 +90,14 @@ Output:
     one floating point number
 '''
 def dSigma_dE(eNu, eE):
-    if abs(get_eE(eNu) - eE) > epsilon:
-        # This should never happen, since we set bounds_eE() accordingly above
+    if eNu < e_thr or abs(get_eE(eNu) - eE) > epsilon:
+        # This should never happen, since we set bounds for eE and eNu accordingly above
         # ... but just in case:
         return 0
 
-    sigma0 = 3.439E-44 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
-    a1, a2, a3 = 10.164, -0.4666, 0.0546
-    sigma = sigma0 * (a1*(eNu-e_thr) + a2*(eNu-e_thr)**2 + a3*(eNu-e_thr)**3)
+    sigma0 = 1.08E-38 * (5.067731E10)**2 # convert cm^2 to MeV^-2, see http://www.wolframalpha.com/input/?i=cm%2F(hbar+*+c)+in+MeV%5E(-1)
+    sigma = sigma0 * ((eNu-e_thr) / 939)**2
+    sigma *= 1.11**2 # coupling constant \beta (see discussion above)
     return sigma / (2*epsilon) # Ensure that integration over eE yields sigma
 
 
@@ -107,13 +111,12 @@ Output:
     one floating point number
 '''
 def dSigma_dCosT(eNu, cosT):
-    # Small values of cosT are preferred, see arXiv:hep-ex/0105068 (fig. 12,14).
-    # However, energy dependence is unclear, so we use a constant value for now.
+    # Energy dependence is unclear, so we use a constant value for now.
     if abs(cosT) > 1:
         return 0
     return 0.5
 
 
-# minimum/maximum neutrino energy that can produce a given positron energy
+# min/max neutrino energy to produce outgoing particle with observed energy
 def _bounds_eNu(eE):
-    return (eE + e_thr - epsilon, eE + e_thr + epsilon)
+    return bounds_eNu
